WebSite/requirements.txt
Flask
flask_migrate
flask_sqlalchemy
python-dotenv
requests
gunicorn